+++ src/SdtdMultiServerKit/Data/Repositories/PointsInfoRepository.cs
@@ -3,7 +3,6 @@ using SdtdMultiServerKit.Data.Entities;
 using SdtdMultiServerKit.Data.IRepositories;
 using SdtdMultiServerKit.Panel;
 using System.Text;
-using Webserver.WebAPI.APIs.WorldState;
 
 namespace SdtdMultiServerKit.Data.Repositories
 {