+++ src/SdtdMultiServerKit/Functions/TeleportFriend.cs
@@ -5,7 +5,6 @@ using SdtdMultiServerKit.Managers;
 using SdtdMultiServerKit.Models;
 using SdtdMultiServerKit.Variables;
 using System.Collections.Concurrent;
-using Webserver.WebAPI.APIs.WorldState;
 
 namespace SdtdMultiServerKit.Functions
 {