+++ src/SdtdMultiServerKit/Functions/TeleportHome.cs
@@ -3,7 +3,6 @@ using SdtdMultiServerKit.Data.IRepositories;
 using SdtdMultiServerKit.FunctionSettings;
 using SdtdMultiServerKit.Managers;
 using SdtdMultiServerKit.Variables;
-using Webserver.WebAPI.APIs.WorldState;
 
 namespace SdtdMultiServerKit.Functions
 {