+++ src/SdtdMultiServerKit/ModEventHub.cs
@@ -3,7 +3,6 @@ using SdtdMultiServerKit.Managers;
 using System.Text;
 using UnityEngine;
 using UnityEngine.Windows;
-using Webserver.WebAPI.APIs.WorldState;
 using static ModEvents;
 
 namespace SdtdMultiServerKit