+++ src/SdtdMultiServerKit/WebApi/Controllers/PlayersController.cs
@@ -1,7 +1,6 @@
 ﻿using SdtdMultiServerKit.Data.IRepositories;
 using SdtdMultiServerKit.Managers;
 using SdtdMultiServerKit.Models;
-using Webserver.WebAPI.APIs.WorldState;
 
 namespace SdtdMultiServerKit.WebApi.Controllers
 {